refactor: extracted TaskRunner types fix: show log message only if there are changes.
