+++ src/task/index.ts
@@ -1,3 +1,4 @@
-export *  from "./TaskRunner"
-export * from "./TaskManager"
-export * from "./Worker"
+export * from "./TaskRunner";
+export * from "./TaskManager";
+export * from "./Worker";
+export * from "./types";
